Revert "Move ValuesNode to SPI"
This reverts commit 6165aeed673cadfd5130a6342f4d953ca989c337.
