+++ presto-main/src/main/java/com/facebook/presto/sql/planner/plan/InternalPlanVisitor.java
@@ -64,6 +64,11 @@ public abstract class InternalPlanVisitor<R, C>
         return visitPlan(node, context);
     }
 
+    public R visitValues(ValuesNode node, C context)
+    {
+        return visitPlan(node, context);
+    }
+
     public R visitIndexSource(IndexSourceNode node, C context)
     {
         return visitPlan(node, context);
