+++ presto-main/src/main/java/com/facebook/presto/sql/planner/plan/Patterns.java
@@ -18,7 +18,6 @@ import com.facebook.presto.matching.Property;
 import com.facebook.presto.spi.plan.FilterNode;
 import com.facebook.presto.spi.plan.PlanNode;
 import com.facebook.presto.spi.plan.TableScanNode;
-import com.facebook.presto.spi.plan.ValuesNode;
 import com.facebook.presto.spi.relation.RowExpression;
 import com.facebook.presto.spi.relation.VariableReferenceExpression;
