+++ presto-main/src/test/java/com/facebook/presto/sql/query/TestSubqueries.java
@@ -13,12 +13,12 @@
  */
 package com.facebook.presto.sql.query;
 
-import com.facebook.presto.spi.plan.ValuesNode;
 import com.facebook.presto.sql.planner.Plan;
 import com.facebook.presto.sql.planner.assertions.PlanMatchPattern;
 import com.facebook.presto.sql.planner.plan.AggregationNode;
 import com.facebook.presto.sql.planner.plan.JoinNode;
 import com.facebook.presto.sql.planner.plan.ProjectNode;
+import com.facebook.presto.sql.planner.plan.ValuesNode;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import org.intellij.lang.annotations.Language;
