+++ presto-spi/src/main/java/com/facebook/presto/spi/plan/PlanVisitor.java
@@ -29,9 +29,4 @@ public abstract class PlanVisitor<R, C>
     {
         return visitPlan(node, context);
     }
-
-    public R visitValues(ValuesNode node, C context)
-    {
-        return visitPlan(node, context);
-    }
 }
